+++ src/ui/side-panel/build-info.js
@@ -2,4 +2,4 @@
  * [IMPL-SIDE_PANEL_TABS] Build-time compile time (UTC). Overwritten by scripts/build.js before side-panel bundle.
  * Format: YYYY-MM-DD HH:mm
  */
-export const BUILD_TIME_UTC = '2026-07-29 05:57'
+export const BUILD_TIME_UTC = '2026-07-29 17:30'